+++ packages/spruce-cli/src/builders/FormBuilder.ts
@@ -10,6 +10,7 @@ import Schema, {
 	ErrorCode as SchemaErrorCode,
 	SchemaError
 } from '@sprucelabs/schema'
+import { IFieldDefinition } from '@sprucelabs/schema'
 import { pick } from 'lodash'
 import { ErrorCode } from '#spruce/errors/codes.types'
 import SpruceError from '../errors/SpruceError'
@@ -177,9 +178,9 @@ export default class FormBuilder<T extends ISchemaDefinition> extends Schema<
 				friendlyMessage: 'Form builder does not support isArray yet'
 			})
 		}
-		// TODO need is array support
+
 		if (value) {
-			definition.defaultValue = value
+			definition.defaultValue = value as IFieldDefinition['defaultValue']
 		}
 
 		// Do we have a lister?