+++ packages/spruce-cli/src/stores/SchemaStore.ts
@@ -28,6 +28,8 @@ import AbstractStore from './AbstractStore'
 
 export interface ISchemaTemplateItemsOptions {
 	includeErrors?: boolean
+	/* Where should i look for local definitions? */
+	localLookupDir: string
 }
 
 export interface IFieldTemplateItemsOptions
@@ -59,9 +61,9 @@ export default class SchemaStore extends AbstractStore {
 
 	/** Get the schema map supplied by core */
 	public async schemaTemplateItems<T extends ISchemaTemplateItemsOptions>(
-		options?: T
+		options: T
 	): Promise<SchemaTemplateItemsReturnType<T>> {
-		const { includeErrors = true } = options ?? {}
+		const { includeErrors = true, localLookupDir } = options
 
 		/** Get all schemas from api  */
 		// TODO load from api
@@ -86,9 +88,7 @@ export default class SchemaStore extends AbstractStore {
 		const localDefinitions = (
 			await Promise.all(
 				(
-					await globby([
-						pathUtil.join(this.cwd, '/src/schemas/**/*.definition.ts')
-					])
+					await globby([pathUtil.join(localLookupDir, '/**/*.definition.ts')])
 				).map(async file => {
 					try {
 						const definition = await this.services.child.importDefault(file, {
@@ -225,12 +225,12 @@ export default class SchemaStore extends AbstractStore {
 			const name = registration.className
 
 			types.push({
-				pascalName: this.utilities.names.toPascal(name),
-				camelName: this.utilities.names.toCamel(name),
+				namePascal: this.utilities.names.toPascal(name),
+				nameCamel: this.utilities.names.toCamel(name),
 				package: pkg,
 				className: registration.className,
 				importAs,
-				readableName: registration.className,
+				nameReadable: registration.className,
 				pascalType: this.utilities.names.toPascal(registration.type),
 				camelType: this.utilities.names.toCamel(registration.type),
 				isLocal: addon.isLocal,